+++ plugins/comparative-adapters/src/index.ts
@@ -3,7 +3,7 @@ import PluginManager from '@jbrowse/core/PluginManager'
 import { FileLocation } from '@jbrowse/core/util/types'
 
 import PAFAdapterF from './PAFAdapter'
-import MCScanAnchorsAdapterF from './PAFAdapter'
+import MCScanAnchorsAdapterF from './MCScanAnchorsAdapter'
 import DeltaAdapterF from './DeltaAdapter'
 import ChainAdapterF from './ChainAdapter'
