+++ plugins/linear-comparative-view/src/LinearSyntenyRenderer/configSchema.ts
