--- plugins/linear-comparative-view/src/LinearSyntenyRenderer/index.js
@@ -1,3 +1,0 @@
-export { default as ReactComponent } from './components/LinearSyntenyRendering'
-export { default as configSchema } from './configSchema'
-export { default } from './LinearSyntenyRenderer'
