+++ plugins/linear-comparative-view/src/LinearSyntenyRenderer/index.ts
@@ -1,0 +1,17 @@
+import PluginManager from '@jbrowse/core/PluginManager'
+
+import ReactComponent from './components/LinearSyntenyRendering'
+import configSchema from './configSchema'
+import LinearSyntenyRenderer from './LinearSyntenyRenderer'
+
+export default (pluginManager: PluginManager) => {
+  pluginManager.addRendererType(
+    () =>
+      new LinearSyntenyRenderer({
+        name: 'LinearSyntenyRenderer',
+        configSchema,
+        ReactComponent,
+        pluginManager,
+      }),
+  )
+}
