+++ plugins/linear-comparative-view/src/index.tsx
@@ -1,10 +1,4 @@
-import { ConfigurationSchema } from '@jbrowse/core/configuration'
 import DisplayType from '@jbrowse/core/pluggableElementTypes/DisplayType'
-import {
-  createBaseTrackConfig,
-  createBaseTrackModel,
-} from '@jbrowse/core/pluggableElementTypes/models'
-import TrackType from '@jbrowse/core/pluggableElementTypes/TrackType'
 import Plugin from '@jbrowse/core/Plugin'
 import PluginManager from '@jbrowse/core/PluginManager'
 import {
@@ -19,85 +13,26 @@ import ViewType from '@jbrowse/core/pluggableElementTypes/ViewType'
 
 import AddIcon from '@material-ui/icons/Add'
 import CalendarIcon from '@material-ui/icons/CalendarViewDay'
-// locals
-//
-import {
-  configSchemaFactory as linearComparativeDisplayConfigSchemaFactory,
-  ReactComponent as LinearComparativeDisplayReactComponent,
-  stateModelFactory as linearComparativeDisplayStateModelFactory,
-} from './LinearComparativeDisplay'
-import LinearComparativeViewFactory from './LinearComparativeView'
-import {
-  configSchemaFactory as linearSyntenyDisplayConfigSchemaFactory,
-  stateModelFactory as linearSyntenyDisplayStateModelFactory,
-} from './LinearSyntenyDisplay'
-import LinearSyntenyRenderer, {
-  configSchema as linearSyntenyRendererConfigSchema,
-  ReactComponent as LinearSyntenyRendererReactComponent,
-} from './LinearSyntenyRenderer'
-import LinearSyntenyViewFactory from './LinearSyntenyView'
+import LinearComparativeDisplayF from './LinearComparativeDisplay'
+import LinearComparativeViewF from './LinearComparativeView'
+import LinearSyntenyDisplayF from './LinearSyntenyDisplay'
+import LinearSyntenyRendererF from './LinearSyntenyRenderer'
+import LinearSyntenyViewF from './LinearSyntenyView'
+import SyntenyTrackF from './SyntenyTrack'
+import { WindowSizeDlg } from './ReadVsRef'
 
 export default class extends Plugin {
   name = 'LinearComparativeViewPlugin'
 
   install(pluginManager: PluginManager) {
     pluginManager.addViewType(() =>
-      pluginManager.jbrequire(LinearComparativeViewFactory),
+      pluginManager.jbrequire(LinearComparativeViewF),
     )
-    pluginManager.addViewType(() =>
-      pluginManager.jbrequire(LinearSyntenyViewFactory),
-    )
-
-    pluginManager.addTrackType(() => {
-      const configSchema = ConfigurationSchema(
-        'SyntenyTrack',
-        {},
-        { baseConfiguration: createBaseTrackConfig(pluginManager) },
-      )
-      return new TrackType({
-        name: 'SyntenyTrack',
-        configSchema,
-        stateModel: createBaseTrackModel(
-          pluginManager,
-          'SyntenyTrack',
-          configSchema,
-        ),
-      })
-    })
-    pluginManager.addDisplayType(() => {
-      const configSchema =
-        linearComparativeDisplayConfigSchemaFactory(pluginManager)
-      return new DisplayType({
-        name: 'LinearComparativeDisplay',
-        configSchema,
-        stateModel: linearComparativeDisplayStateModelFactory(configSchema),
-        trackType: 'SyntenyTrack',
-        viewType: 'LinearComparativeView',
-        ReactComponent: LinearComparativeDisplayReactComponent,
-      })
-    })
-    pluginManager.addDisplayType(() => {
-      const configSchema =
-        linearSyntenyDisplayConfigSchemaFactory(pluginManager)
-      return new DisplayType({
-        name: 'LinearSyntenyDisplay',
-        configSchema,
-        stateModel: linearSyntenyDisplayStateModelFactory(configSchema),
-        trackType: 'SyntenyTrack',
-        viewType: 'LinearSyntenyView',
-        ReactComponent: LinearComparativeDisplayReactComponent,
-      })
-    })
-
-    pluginManager.addRendererType(
-      () =>
-        new LinearSyntenyRenderer({
-          name: 'LinearSyntenyRenderer',
-          configSchema: linearSyntenyRendererConfigSchema,
-          ReactComponent: LinearSyntenyRendererReactComponent,
-          pluginManager,
-        }),
-    )
+    pluginManager.addViewType(() => pluginManager.jbrequire(LinearSyntenyViewF))
+    LinearSyntenyRendererF(pluginManager)
+    LinearComparativeDisplayF(pluginManager)
+    LinearSyntenyDisplayF(pluginManager)
+    SyntenyTrackF(pluginManager)
 
     pluginManager.addToExtensionPoint(
       'Core-extendPluggableElement',
@@ -120,16 +55,14 @@ export default class extends Plugin {
                         label: 'Linear read vs ref',
                         icon: AddIcon,
                         onClick: () => {
-                          getSession(self).queueDialog(
-                            (doneCallback: Function) => [
-                              WindowSizeDlg,
-                              {
-                                track: getContainingTrack(self),
-                                feature,
-                                handleClose: doneCallback,
-                              },
-                            ],
-                          )
+                          getSession(self).queueDialog(doneCallback => [
+                            WindowSizeDlg,
+                            {
+                              track: getContainingTrack(self),
+                              feature,
+                              handleClose: doneCallback,
+                            },
+                          ])
                         },
                       },
                     ]
